chore(mobile): Factor out ItemSelector from ItemSelectorModal to use directly in components
